Backend/requirements.txt
flask
praw
openai
pyyaml
flasgger
python-dotenv
psycopg2-binary
sqlalchemy